testipc to be tested with gdb

+++ TP-Arqui/Kernel/processes/process.c
@@ -358,7 +358,7 @@ process_control_block *get_PCB(processManagerADT pm, uint64_t pid) {
   return pm->process_table[pid].pcb;
 }
 
-int8_t set_fd(processManagerADT pm, uint8_t fd, uint8_t pipe_id) {
+int8_t set_fd(processManagerADT pm, uint16_t fd, uint16_t pipe_id) {
   process_control_block *pcb = get_running(pm->scheduler);
   if (fd >= FD_COUNT || pipe_id >= MAX_PIPES_COUNT ||
       !check_pipe_id(my_pipe_manager, pipe_id)) {
